+++ internal/application/points_calculator.go
@@ -1,1 +1,2 @@
 // Logic for calculating points based on receipt rules
+package application
