+++ internal/application/receipt_rules.go
@@ -1,1 +1,2 @@
 // Encapsulates the rules for calculating points
+package application
